[ISSUE-12] DRY SharedZookeeperTestResource between Junit4 and Junit5 implementations
